fix: Detect the changed files using git
BREAKING CHANGE: Instead of adding changelog and package descriptor to the changed files, git is used to detect the changed files. Although there probably are not other changed files in the local repo after running this action, there might be in some build configurations. In suich case, this action should be run as the first one, to collect only changes made by the version number bumping.